add braintrust parent parsing util
